--- client/src/app/calendar/_components/CalendarWrapper.tsx
@@ -1,11 +1,0 @@
-'use client';
-
-import ViewEvents from './ViewEvents';
-
-export default function CalendarWrapper() {
-  return (
-    <>
-      <ViewEvents />
-    </>
-  );
-}
