+++ server/src/email/resend.ts
@@ -1,8 +1,9 @@
+import { Fragment, createElement } from 'react';
 import { CreateEmailOptions, Resend } from 'resend';
+import { render } from '@react-email/components';
 
 import { Senders } from './senders';
 import { isDev } from '@@/util/dev';
-import { render } from '@react-email/components';
 import { brevo } from '.';
 
 const resend = new Resend(process.env.RESEND_API_KEY ?? 'DEVELOPMENT');
@@ -15,7 +16,7 @@ export async function send(
 
   const { error } = await resend.emails.send({
     ...props,
-    text: props.text ?? (await render(<>{props.react}</>, { plainText: true })),
+    text: props.text ?? (await render(el(props.react), { plainText: true })),
   });
   if (error) {
     if (error.name.includes('invalid') || error.name.includes('missing'))
@@ -30,7 +31,7 @@ export async function send(
 
 function localSend(content: React.ReactNode, title?: string) {
   return catchTF(async () => {
-    const text = await render(<>{content}</>, { plainText: true });
+    const text = await render(el(content), { plainText: true });
     console.log(`${title || 'EMAIL CONTENTS'}\n`, text);
   });
 }
@@ -41,8 +42,8 @@ function brevoFallback(
   content: React.ReactNode
 ) {
   return catchTF(async () => {
-    const html = await render(<>{content}</>);
-    const text = await render(<>{content}</>, { plainText: true });
+    const html = await render(el(content));
+    const text = await render(el(content), { plainText: true });
 
     await brevo.sendRawEmail({ to, subject, html, text });
   });
@@ -57,4 +58,8 @@ async function catchTF(cb: (...p: any[]) => any): Promise<boolean> {
   } catch (_) {
     return false;
   }
+}
+
+function el(content: React.ReactNode) {
+  return createElement(Fragment, {}, content);
 }
